requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
config==0.4.0
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.2.0
idna==2.7
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
line-profiler==2.1.2
MarkupSafe==1.1.0
matplotlib==2.2.3
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
numpy==1.15.3
pandas==0.23.4
pandocfilters==1.4.2
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.3.0
pyparsing==2.3.0
python-dateutil==2.7.3
python-snappy==0.5.3
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.3
redis==2.10.6
requests==2.20.1
scandir==1.9.0
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
subprocess32==3.5.3
terminado==0.8.1
testpath==0.4.2
torch==1.0.0
torchvision==0.2.1
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xgboost==0.81
